Modificacion de inscripciones a controller
