+++ data/courses.json
@@ -22,8 +22,7 @@
     "area": "programacion",
     "estado": "activa",
     "alumnos": [
-      1,
-      2
+      3
     ]
   },
   {
@@ -162,7 +161,6 @@
     "estado": "activa",
     "alumnos": [
       "1",
-      "2",
       "3",
       "4",
       "5",
@@ -170,7 +168,8 @@
       "7",
       "8",
       3,
-      4
+      4,
+      6
     ]
   },
   {
